feat: refactor loan form and wizard controllers, separate logic and enhance error handling
